backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.6.1
google-generativeai==0.8.3
google-auth==2.36.0
httpx==0.28.1
python-multipart==0.0.12
python-dotenv==1.0.1
aiofiles==24.1.0
